requirements.txt
tomli
tomli_w
# nvdlib
jsonlib-python3